+++ filament/include/filament/View.h
@@ -159,6 +159,16 @@ public:
         bool threshold = true;                  //!< whether to threshold the source
         bool enabled = false;                   //!< enable or disable bloom
         float highlight = 1000.0f;              //!< limit highlights to this value before bloom [10, +inf]
+
+        bool lensFlare = false;                 //!< enable screen-space lens flare
+        bool starburst = true;                  //!< enable starburst effect on lens flare
+        float chromaticAberration = 0.005f;     //!< amount of chromatic aberration
+        uint8_t ghostCount = 4;                 //!< number of flare "ghosts"
+        float ghostSpacing = 0.6f;              //!< spacing of the ghost in screen units [0, 1[
+        float ghostThreshold = 10.0f;           //!< hdr threshold for the ghosts
+        float haloThickness = 0.1f;             //!< thickness of halo in vertical screen units, 0 to disable
+        float haloRadius = 0.4f;                //!< radius of halo in vertical screen units [0, 0.5]
+        float haloThreshold = 10.0f;            //!< hdr threshold for the halo
     };
 
     /**