+++ RELEASE_NOTES.md
@@ -5,6 +5,8 @@ A new header is inserted each time a *tag* is created.
 
 ## Next release (main branch)
 
+- engine: new screen-space lens flare
+
 ## v1.9.24
 
 ## v1.9.23